+++ SimulationDataProducts/EnergyDeposit.h
@@ -11,16 +11,16 @@
 
 namespace gar {
   namespace sdp {
-    
-    
+
+
     class EnergyDeposit{
     public:
-      
+
       EnergyDeposit();
-      
+
 #ifndef __GCCXML__
       EnergyDeposit(int    trackID,
-                    double t,
+                    float t,
                     float  e,
                     float  xpos,
                     float  ypos,
@@ -36,24 +36,24 @@ namespace gar {
       , fdX       (length)
       , fIsPrimary(isPrimary)
       {}
-      
+
       int    const& TrackID()   const { return fTrackID;   }
-      double const& Time()      const { return fTime;      }
+      float  const& Time()      const { return fTime;      }
       float  const& Energy()    const { return fEnergy;    }
       float  const& X()         const { return fX;         }
       float  const& Y()         const { return fY;         }
       float  const& Z()         const { return fZ;         }
       float  const& dX()        const { return fdX;        }
       bool          IsPrimary() const { return fIsPrimary; }
-      
+
       bool operator  <(gar::sdp::EnergyDeposit const& b) const;
-      
+
 #endif
-      
+
     private:
-      
+
       int    fTrackID;   ///< g4 track ID of particle making the deposit
-      double fTime;      ///< time of the energy deposit
+      float fTime;      ///< time of the energy deposit
       float  fEnergy;    ///< energy deposited
       float  fX;         ///< x position of the energy deposit
       float  fY;         ///< y position of the energy deposit
@@ -61,7 +61,7 @@ namespace gar {
       float  fdX;        ///< the size of the step for this energy deposit
       bool   fIsPrimary; ///< is this from the primary particle, or a secondary (ie EM)
     };
-    
+
 
   } // sdp
 } // gar
